Switch to scraping the 6th Legislative Council
There was an election in September and the 5th Legislative Council is
over; the scraper also no longer worked because since then the pages
for the 5th term have been changed to remove links to the people
pages.

This commit changes the scraper to scrape the 6th term pages instead.